// src/pages/WorkInProgressPage/types/workInProgressTypes.ts
export interface ITableRow {
    key: string;
    meltNumber: string;
    packageNumber: number;
    date: string;
    status: string;
    length: number;
    specification: string;
    brand: string;
    certificateNumber: string;
    section: number;
    controlScheme: string;
    material: string;
    net: number;
    gross: number;
  }